A22DAMPIMOBI-2098 : [Điểm danh] sửa lại theo ghép api điểm danh khuôn mặt

See merge request quannv_edsolabs/ccvc-mobile!2106

+++ lib/diem_danh_module/data/response/thong_ke_diem_danh_ca_nhan_response.dart
@@ -27,7 +27,7 @@ class ThongKeDiemDanhCaNhanModelResponse {
   @JsonKey(name: 'soLanVeSom')
   int? soLanVeSom;
   @JsonKey(name: 'soNgayLamViec')
-  int? soNgayLamViec;
+  double? soNgayLamViec;
   @JsonKey(name: 'soLanChamCongThuCong')
   int? soLanChamCongThuCong;
   @JsonKey(name: 'soNgayVangMatKhongLyDo')
